# docs/api-contract.yaml
openapi: 3.0.3
info:
  title: VitaCare API
  description: |
    VitaCare Caregiver Onboarding Platform API.

    ## Authentication
    All endpoints (except those under `/auth`) require a valid JWT Bearer token
    in the `Authorization` header.

    ## Response Format
    All responses follow a standard envelope:
    - **Success:** `{ success: true, data: {...}, meta?: {...} }`
    - **Error:** `{ success: false, error: { code: "XXX_NNN", message: "..." } }`
  version: "1.0.0"
  contact:
    name: VitaCare Engineering
    email: vitacasahealthindia@gmail.com
servers:
  - url: https://api.vitacasahealth.in/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Custom JWT issued by the VitaCare API. Access token TTL is 1 hour.
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "AUTH_001"
            message:
              type: string
              example: "Phone number is already registered"
...
    Gender:
      type: string
      enum: [male, female, other]
    Language:
      type: string
      enum: [hindi, english, kannada, tamil, telugu, malayalam, bengali, gujarati, marathi]
    ServiceMode:
      type: string
      enum: [24hrs_live_in, 12hrs_pg]
    Religion:
      type: string
      enum: [hindu, muslim, christian, others]
...
    City:
      type: string
      enum: [bangalore, mumbai, hyderabad, chennai, pune, delhi, gurgaon]
    Qualification:
      type: string
      enum:
        - bsc_gnm_completed
        - anm_completed
        - bsc_gnm_anm_backlog
        - bsc_gnm_anm_student
        - non_nursing
    VerificationStatus:
      type: string
      enum:
        - pending_call
        - call_verified
        - pending_verification
        - in_process
        - available
        - unavailable
        - assigned
        - rejected
    DocumentType:
      type: string
      enum: [qualification, aadhaar, other]
...
    CaregiverProfile:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
        profile_id:
          type: string
          format: uuid
        full_name:
          type: string
          example: "Ramesh Kumar"
        phone:
          type: string
          example: "+919876543210"
        email:
          type: string
          format: email
          nullable: true
          example: "ramesh@example.com"
        gender:
          $ref: "#/components/schemas/Gender"
        age:
          type: integer
          minimum: 18
          maximum: 65
          example: 32
        selfie_photo_url:
          type: string
          nullable: true
          description: Signed URL (1-hour expiry)
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
        service_modes:
          type: array
          items:
            $ref: "#/components/schemas/ServiceMode"
          nullable: true
        highest_qualification:
          $ref: "#/components/schemas/Qualification"
        qualification_document_url:
          type: string
          nullable: true
          description: Signed URL (1-hour expiry)
        aadhaar_document_url:
          type: string
          nullable: true
          description: Signed URL (1-hour expiry)
        other_document_urls:
          type: array
          items:
            type: string
          nullable: true
          description: Array of signed URLs (1-hour expiry)
        current_address:
          type: string
          nullable: true
          maxLength: 500
        terms_accepted:
          type: boolean
        verification_status:
          $ref: "#/components/schemas/VerificationStatus"
        rejection_message:
          type: string
          nullable: true
        advanced_details_completed:
          type: boolean
        preferred_cities:
          type: array
          items:
            $ref: "#/components/schemas/City"
          description: Multi-select — zero or more preferred cities.
        notes:
          type: string
          nullable: true
          maxLength: 500
        created_at:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: "AUTH_005"
              message: "Invalid or expired token"
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: "GEN_001"
              message: "Invalid request body"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error:
              code: "GEN_003"
              message: "Internal server error"
paths:
...
  /caregiver/profile:
    get:
      tags: [Caregiver Profile]
      summary: Get own profile
      description: Returns the full caregiver profile with signed document URLs. All fields are always returned; unset fields return null.
      operationId: getCaregiverProfile
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Profile retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/CaregiverProfile"
              example:
                success: true
                data:
                  user_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                  profile_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                  full_name: "Ramesh Kumar"
                  phone: "+919876543210"
                  email: "ramesh@example.com"
                  gender: "male"
                  age: 32
                  selfie_photo_url: "https://storage.supabase.co/signed/..."
                  languages: ["hindi", "english"]
                  service_modes: ["24hrs_live_in", "12hrs_pg"]
                  highest_qualification: "bsc_gnm_completed"
                  qualification_document_url: "https://storage.supabase.co/signed/..."
                  aadhaar_document_url: "https://storage.supabase.co/signed/..."
                  other_document_urls: ["https://storage.supabase.co/signed/..."]
                  current_address: "123, MG Road, Bangalore"
                  terms_accepted: true
                  verification_status: "available"
                  rejection_message: null
                  advanced_details_completed: true
                  preferred_cities: ["bangalore", "mumbai"]
                  notes: "Available for night shifts"
                  created_at: "2026-08-01T10:00:00Z"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /caregiver/profile/basic:
    put:
      tags: [Caregiver Profile]
      summary: Update basic profile fields
      description: |
        Updates basic profile information. Does NOT change verification status.
        Profile is flagged as `has_pending_edits = true` for admin review.
        Does NOT include `full_name` or `gender` — caregivers cannot self-edit
        their name or gender once set at registration; only admins can
        (PUT /admin/caregivers/{id}).
      operationId: updateBasicProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [age, languages]
              properties:
                age:
                  type: integer
                  minimum: 18
                  maximum: 65
                  example: 33
                languages:
                  type: array
                  minItems: 1
                  items:
                    $ref: "#/components/schemas/Language"
                  example: ["hindi", "english", "kannada"]
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      has_pending_edits:
                        type: boolean
                      verification_status:
                        $ref: "#/components/schemas/VerificationStatus"
              example:
                success: true
                data:
                  message: "Profile updated"
                  has_pending_edits: true
                  verification_status: "available"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /caregiver/profile/advanced:
    put:
      tags: [Caregiver Profile]
      summary: Submit advanced details
      description: |
        Submit advanced profile details. Only accessible when verification_status is `call_verified`.
        Requires the Aadhaar document to already be uploaded (selfie is already guaranteed by
        registration; qualification and "other" documents are optional, not required).
        On success, status transitions to `pending_verification`.
      operationId: submitAdvancedDetails
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [highest_qualification, religion, terms_accepted]
              properties:
                highest_qualification:
                  $ref: "#/components/schemas/Qualification"
                religion:
                  $ref: "#/components/schemas/Religion"
                  description: "Required. Caregiver's religion."
                father_name:
                  type: string
                  nullable: true
                  minLength: 1
                  maxLength: 100
                  pattern: "^[a-zA-Z\\s]+$"
                  description: "Optional. Father's full name. Alphabetic + spaces only if provided."
                  example: "Suresh Kumar"
                father_phone:
                  type: string
                  nullable: true
                  pattern: "^\\+91[6-9]\\d{9}$"
                  description: "Optional. Father's phone number in +91XXXXXXXXXX format if provided."
                  example: "+919876500001"
                current_address:
                  type: string
                  nullable: true
                  minLength: 1
                  maxLength: 500
                  description: "Optional. Max 500 characters if provided."
                  example: "123, MG Road, Bangalore 560001"
                preferred_cities:
                  type: array
                  items:
                    $ref: "#/components/schemas/City"
                  description: "Multi-select. Optional — omitted or [] means no preference."
                  example: ["bangalore", "mumbai"]
                notes:
                  type: string
                  nullable: true
                  maxLength: 500
                  description: "Optional notes, max 500 characters"
                  example: "Available for night shifts"
                terms_accepted:
                  type: boolean
                  description: "Must be true"
                  enum: [true]
      responses:
        "200":
          description: Advanced details submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      verification_status:
                        type: string
                        enum: [pending_verification]
              example:
                success: true
                data:
                  message: "Advanced details submitted"
                  verification_status: "pending_verification"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missing_docs:
                  summary: Aadhaar document not uploaded
                  value:
                    success: false
                    error:
                      code: "PROFILE_017"
                      message: "Aadhaar card not uploaded. Please upload it before submitting."
                terms_not_accepted:
                  summary: Terms not accepted
                  value:
                    success: false
                    error:
                      code: "PROFILE_009"
                      message: "Terms and conditions must be accepted"
                unexpected_code_field:
                  summary: A `code` field was included (login code is set at registration, not here)
                  value:
                    success: false
                    error:
                      code: "GEN_001"
                      message: "Invalid request body"
        "403":
          description: Not in call_verified status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "PROFILE_008"
                  message: "Advanced details not available. Phone call verification required."
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "AUTH_008"
                  message: "Email is already registered"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      tags: [Caregiver Profile]
      summary: Edit advanced profile fields
      description: |
        Self-edit of the advanced-details fields any time after the initial
        submission — unlike the PUT above (one-time/resubmission-only, whole-object,
        resets status to `pending_verification`), this is a partial update
        (any subset of fields) that never touches verification_status, only
        flags `has_pending_edits = true`. Requires `advanced_details_completed = true`.
        Does NOT include `religion` — once set (via the PUT above, on initial
        submission or a rejected resubmission), it's locked from self-edit;
        only admins can change it (PUT /admin/caregivers/{id}).
      operationId: editAdvancedProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Any subset of these fields — only what's provided gets written.
              properties:
                highest_qualification:
                  $ref: "#/components/schemas/Qualification"
                father_name:
                  type: string
                  nullable: true
                  pattern: "^[a-zA-Z\\s]+$"
                father_phone:
                  type: string
                  nullable: true
                  pattern: "^\\+91[6-9]\\d{9}$"
                current_address:
                  type: string
                  maxLength: 500
                preferred_cities:
                  type: array
                  items:
                    $ref: "#/components/schemas/City"
                notes:
                  type: string
                  nullable: true
                  maxLength: 500
            example:
              current_address: "456, New Address, Bangalore 560001"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      has_pending_edits:
                        type: boolean
              example:
                success: true
                data:
                  message: "Profile updated"
                  has_pending_edits: true
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          description: Advanced details never submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "PROFILE_025"
                  message: "Advanced details must be submitted before editing them"
        "500":
          $ref: "#/components/responses/InternalError"
  /caregiver/profile/phone:
    patch:
      tags: [Caregiver Profile]
      summary: Change phone number
      description: |
        Changes the caregiver's own login phone number. Identity-sensitive:
        if the caregiver is currently `available` or `unavailable`, this
        resets them to `pending_verification` for re-review (never from
        `in_process`/`assigned`). For every other status, it just updates
        the phone and flags `has_pending_edits = true`.
      operationId: updatePhone
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone]
              properties:
                phone:
                  type: string
                  pattern: "^\\+91[6-9]\\d{9}$"
                  example: "+919876500099"
      responses:
        "200":
          description: Phone number updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      verification_status:
                        $ref: "#/components/schemas/VerificationStatus"
              example:
                success: true
                data:
                  message: "Phone number updated"
                  verification_status: "pending_verification"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          description: Phone number already registered to another account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "AUTH_001"
                  message: "Phone number is already registered"
        "500":
          $ref: "#/components/responses/InternalError"
  /caregiver/profile/code:
    patch:
      tags: [Caregiver Profile]
      summary: Change login code (PIN)
      description: |
        Changes the caregiver's own 4-digit login code. Never triggers
        re-review — this is an account-security action, not an identity
        change (unlike phone/Aadhaar).
      operationId: updateCode
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: "^\\d{4}$"
                  example: "5678"
      responses:
        "200":
          description: Login code updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
              example:
                success: true
                data:
                  message: "Login code updated"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /caregiver/profile/selfie:
    post:
      tags: [Caregiver Profile]
      summary: Upload selfie photo
      description: Upload selfie via multipart form data. Max 10MB. No file type restriction. Camera capture only (enforced client-side).
      operationId: uploadSelfie
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: "Selfie image file. Max 10MB."
      responses:
        "200":
          description: Selfie uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      file_path:
                        type: string
              example:
                success: true
                data:
                  message: "Selfie uploaded"
                  file_path: "caregiver-documents/b2c3d4e5-f6a7-8901-bcde-f12345678901/selfie.jpg"
        "400":
          description: Upload validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                no_file:
                  summary: No file provided
                  value:
                    success: false
                    error:
                      code: "UPLOAD_001"
                      message: "File is required"
                too_large:
                  summary: File exceeds size limit
                  value:
                    success: false
                    error:
                      code: "UPLOAD_002"
                      message: "File size exceeds 10MB limit"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          description: Upload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "UPLOAD_005"
                  message: "File upload failed. Please try again."
  /caregiver/profile/documents:
    post:
      tags: [Caregiver Profile]
      summary: Upload documents
      description: |
        Upload qualification, Aadhaar, or other documents via multipart form data.
        - `qualification` and `aadhaar`: One file each, re-upload overwrites.
        - `other`: Up to 3 files maximum.
        Max 10MB per file. No file type restriction.
      operationId: uploadDocument
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, document_type]
              properties:
                file:
                  type: string
                  format: binary
                  description: "Document file. Max 10MB."
                document_type:
                  $ref: "#/components/schemas/DocumentType"
      responses:
        "200":
          description: Document uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                      document_type:
                        $ref: "#/components/schemas/DocumentType"
                      file_path:
                        type: string
              example:
                success: true
                data:
                  message: "Document uploaded"
                  document_type: "aadhaar"
                  file_path: "caregiver-documents/b2c3d4e5-f6a7-8901-bcde-f12345678901/aadhaar.pdf"
        "400":
          description: Upload validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                no_file:
                  summary: No file provided
                  value:
                    success: false
                    error:
                      code: "UPLOAD_001"
                      message: "File is required"
                too_large:
                  summary: File exceeds size limit
                  value:
                    success: false
                    error:
                      code: "UPLOAD_002"
                      message: "File size exceeds 10MB limit"
                max_other_docs:
                  summary: Too many other documents
                  value:
                    success: false
                    error:
                      code: "UPLOAD_003"
                      message: "Maximum 3 additional documents allowed"
                invalid_type:
                  summary: Invalid document type
                  value:
                    success: false
                    error:
                      code: "UPLOAD_004"
                      message: "Invalid document_type. Must be: qualification, aadhaar, or other"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          description: Upload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "UPLOAD_005"
                  message: "File upload failed. Please try again."
  /caregiver/verification-status:
    get:
      tags: [Caregiver Profile]
      summary: Get verification status
      description: Returns current verification status, rejection message, and key timestamps.
      operationId: getVerificationStatus
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      verification_status:
                        $ref: "#/components/schemas/VerificationStatus"
                      rejection_message:
                        type: string
                        nullable: true
                      submitted_at:
                        type: string
                        format: date-time
                        nullable: true
                      verified_at:
                        type: string
                        format: date-time
                        nullable: true
              example:
                success: true
                data:
                  verification_status: "pending_verification"
                  rejection_message: null
                  submitted_at: "2026-08-01T10:00:00Z"
                  verified_at: null
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /caregiver/fcm-token:
    put:
      tags: [Caregiver Profile]
      summary: Update FCM token
      description: Update the Firebase Cloud Messaging token for push notifications. Call after registration, on each app launch, and when token refreshes.
      operationId: updateFcmToken
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                  description: "Firebase Cloud Messaging device token"
                  example: "dK8xM2nQR_-abc123..."
      responses:
        "200":
          description: Token updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      message:
                        type: string
              example:
                success: true
                data:
                  message: "FCM token updated"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: "PROFILE_021"
                  message: "FCM token is required"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ============================================================
  # CAREGIVER JOBS ENDPOINTS (SPEC.md 6.6)
  # ============================================================
...